# PERF: Series.str.get_dummies for "string[pyarrow]" and "string[pyarrow_numpy]" dtypes