- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

```
from asv_bench.benchmarks.strings import Dummies

dtype = "string[pyarrow]"
b = Dummies()
b.setup(dtype)

%timeit b.time_get_dummies(dtype)

# 845 ms ± 51.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)     -> main
# 85.5 ms ± 4.44 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR
```